Delete README.md
I intented to use export_summs function from jtools Library to export the results of a linear regression as a table in a word file format. The issue I got is "Column 6 must be named". The original data frame has 5 columns so I suspect the function generates a new table with 6 columns. The backtrace indicates to use .name_repair to specify repair. I tried but failed to go through tibble. Thank you very much for any help in solving this issue.